backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
firebase-admin==6.2.0
google-cloud-storage==2.13.0
numpy
opencv-python-headless
python-multipart==0.0.6
google-genai==0.3.0
requests==2.31.0
mediapipe==0.10.14
# ── Stage 3-5: GPU CV Pipeline ──
ultralytics>=8.1.0
